+++ CHANGELOG.md
@@ -1,3 +1,9 @@
+v1.2:
+- Support for  HPLTv3 documents.
+- New feature : Domain labels (https://huggingface.co/nvidia/multilingual-domain-classifier)
+- Register Labels are reused when they already exist in the dataset (only for HPLT v3 documents)
+- Bump requeriments.
+
 v1.1:
 - Sample extraction is now integrated in stats processing, and samples now are included in resulting yaml files.
 - Added support for Fineweb, Nemotron-CC and MADLAD datasets.